Version 0.2: Support logging in from 'Permission Denied' page
This is what you get if you click on a link into a gerrit review that is
protected behind authentication. The previous version of this script
only accounted for landing on the home page, detecting logged out state
by the presence of a 'Sign In' link.

We now first click a 'Continue' button if present, then try to log in
(clicking the 'Sign In' link, even from javascript, when the 'Permission
Denied' overlay is present, has no effct).